--- mocks/meets.json
@@ -1,322 +1,0 @@
-[
-    {
-        "id": "KOY",
-        "name": "KILCOY",
-        "location": "QLD",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "BEN",
-        "name": "BENDIGO",
-        "location": "VIC",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "SAC",
-        "name": "SAPPHIRE COAST",
-        "location": "NSW",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "PTA",
-        "name": "PORT AUGUSTA",
-        "location": "SA",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "BDA",
-        "name": "BAIRNSDALE",
-        "location": "VIC",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "NTM",
-        "name": "NORTHAM",
-        "location": "WA",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "DEV",
-        "name": "DEVONPORT",
-        "location": "TAS",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "ALC",
-        "name": "ALICE SPRINGS",
-        "location": "NT",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "SHA",
-        "name": "SHA TIN",
-        "location": "HKG",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "SEO",
-        "name": "SEOUL",
-        "location": "KOR",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "HAN",
-        "name": "HANSHIN",
-        "location": "JPN",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "SCV",
-        "name": "SCOTTSVILLE",
-        "location": "ZAF",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "HPR",
-        "name": "HAMILTON PARK",
-        "location": "GBR",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "SBY",
-        "name": "SALISBURY",
-        "location": "GBR",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "NMT",
-        "name": "NEWMARKET UK",
-        "location": "GBR",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "SLI",
-        "name": "SLIGO",
-        "location": "IRL",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "LGP",
-        "name": "LONGCHAMP",
-        "location": "FRA",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "MUN",
-        "name": "MUNCHEN",
-        "location": "DEU",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "HNO",
-        "name": "HANNOVER",
-        "location": "DEU",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "JAG",
-        "name": "JAGERSRO",
-        "location": "SWE",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "LZG",
-        "name": "LEIPZIG",
-        "location": "DEU",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "PRK",
-        "name": "PARKES",
-        "location": "NSW",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "SWN",
-        "name": "SWAN HILL",
-        "location": "VIC",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "IPS",
-        "name": "IPSWICH",
-        "location": "QLD",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "GRA",
-        "name": "GRAFTON",
-        "location": "NSW",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "SAN",
-        "name": "SANDOWN PARK",
-        "location": "VIC",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "HOB",
-        "name": "HOBART",
-        "location": "TAS",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "BGM",
-        "name": "BUSAN KOREA",
-        "location": "KOR",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "MBR",
-        "name": "MURRAY BRIDGE",
-        "location": "SA",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "MRD",
-        "name": "MANDURAH",
-        "location": "WA",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "SAL",
-        "name": "SALE",
-        "location": "VIC",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "CPL",
-        "name": "CAPALABA",
-        "location": "QLD",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "MTG",
-        "name": "MOUNT GAMBIER",
-        "location": "SA",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "DAR",
-        "name": "DARWIN",
-        "location": "NT",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "HYS",
-        "name": "HYERES",
-        "location": "FRA",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "GGS",
-        "name": "GRAIGNES",
-        "location": "FRA",
-        "date": "2022-05-01"
-    },
-    {
-        "name": "Pelaw Grange (GBR)",
-        "location": "GBR",
-        "date": "2022-05-01"
-    },
-    {
-        "name": "Sunderland (GBR)",
-        "location": "GBR",
-        "date": "2022-05-01"
-    },
-    {
-        "name": "Towcester (GBR)",
-        "location": "GBR",
-        "date": "2022-05-01"
-    },
-    {
-        "name": "Kinsley (GBR)",
-        "location": "GBR",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "BAT",
-        "name": "BATHURST",
-        "location": "NSW",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "WGT",
-        "name": "WINGATUI",
-        "location": "NZL",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "RMP",
-        "name": "REMINGTON PARK",
-        "location": "USA",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "FNP",
-        "name": "FONNER PARK",
-        "location": "USA",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "HTH",
-        "name": "HAWTHORNE",
-        "location": "USA",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "GGF",
-        "name": "GOLDEN GATE",
-        "location": "USA",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "MWS",
-        "name": "MEADOWLANDS",
-        "location": "USA",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "HOO",
-        "name": "HOOSIER PARK",
-        "location": "USA",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "ANI",
-        "name": "SANTA ANITA PARK",
-        "location": "USA",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "NFP",
-        "name": "NORTHFIELD PARK",
-        "location": "USA",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "RGI",
-        "name": "RANGIORA",
-        "location": "NZL",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "RHS",
-        "name": "RICHMOND STRAIGHT",
-        "location": "NSW",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "HSV",
-        "name": "HEALESVILLE",
-        "location": "VIC",
-        "date": "2022-05-01"
-    },
-    {
-        "id": "MAA",
-        "name": "MANUKAU",
-        "location": "NZL",
-        "date": "2022-05-01"
-    }
-]
