--- mocks/stats.json
@@ -1,11 +1,0 @@
-[
-    { 
-        "name": "Total Liquidity", "stat": 10000
-    },
-    { 
-        "name": "In Play", "stat": 100
-    },
-    { 
-        "name": "Performance", "stat": 329.36
-    }
-]
