--- scripts/create_sig.py
@@ -1,2 +1,0 @@
-track = input("Enter Venue [bne]:")
-race_number = input("Race Number:")
