+++ src/components/Buttons/ClaimTokensButton.tsx
@@ -14,7 +14,7 @@ export const ClaimTokensButton: React.FC<Props> = ({
 }) => (
   <button
     onClick={onClick}
-    className="px-5 h-16 bg-gray-100 hover:bg-gray-200 rounded-md border border-gray-500 shadow-md "
+    className="px-5 h-16 bg-gray-100 hover:bg-gray-200 rounded-md border border-gray-500 shadow-md"
   >
     {isLoading ? <Loader /> : `Claim ${tokenName}`}
   </button>
