+++ src/icons/index.ts
@@ -1,2 +1,3 @@
 export { ReactComponent as MetaMaskIcon } from "./metamask.svg";
 export { ReactComponent as WalletConnectIcon } from "./walletConnect.svg";
+export { default as TokenomicsChart } from "./chart.svg";
