--- src/images/index.ts
@@ -1,1 +1,0 @@
-export { default as TokenomicsChart } from "./chart.svg";
